Refactor restriction case reception and assignment screen
