+++ apps/judicial-system/web/messages/index.ts
@@ -10,7 +10,6 @@ export * from './RestrictionCases/Court/courtRecordForm'
 export * from './RestrictionCases/Court/ruling'
 export * from './RestrictionCases/Court/confirmation'
 export * from './RestrictionCases/Court/overview'
-export * from './RestrictionCases/Court/receptionAndAssignment'
 export * from './InvestigationCases/Prosecutor/defendantForm'
 export * from './InvestigationCases/Prosecutor/requestedHearingArrangements'
 export * from './InvestigationCases/Prosecutor/demandsForm'
